+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 0.1
 
 * #1 Support for provider methods in modules
+* #3 Allow binding of generic types
 * #9 Module superclass with noop implementation of Module.configure(..)